+++ src/container-with-most-water/main.cpp
@@ -1,11 +1,10 @@
 #include <iostream>
 #include <vector>
 
-#include "solution.hpp"
 #include "../utils.hpp"
+#include "solution.hpp"
 
 using namespace std;
-
 
 int main()
 {
@@ -20,7 +19,6 @@ int main()
     int result = s.maxArea(height);
 
     cout << "Output: " << result << endl;
-
 
     return 0;
 }
