+++ src/two-sum/main.cpp
@@ -1,11 +1,10 @@
 #include <iostream>
 #include <vector>
 
-#include "solution.hpp"
 #include "../utils.hpp"
+#include "solution.hpp"
 
 using namespace std;
-
 
 int main()
 {
